Refactor + new component BallotCategory
